fix: type error

+++ wrappers/ocr/ocr.go
@@ -17,41 +17,41 @@ type OCRRequest struct {
 }
 
 type OCRWord struct {
-    WordText string  `json:"WordText"`
-    Left     float64 `json:"Left"`
-    Top      float64 `json:"Top"`
-    Height   float64 `json:"Height"`
-    Width    float64 `json:"Width"`
+	WordText string `json:"WordText"`
+	Left     int    `json:"Left"`
+	Top      int    `json:"Top"`
+	Height   int    `json:"Height"`
+	Width    int    `json:"Width"`
 }
 
 type OCRLine struct {
-    LineText  string     `json:"LineText"`
-    Words     []OCRWord  `json:"Words"`
-    MaxHeight float64    `json:"MaxHeight"`
-    MinTop    float64    `json:"MinTop"`
+	LineText  string    `json:"LineText"`
+	Words     []OCRWord `json:"Words"`
+	MaxHeight int       `json:"MaxHeight"`
+	MinTop    int       `json:"MinTop"`
 }
 
 type TextOverlay struct {
-    Lines       []OCRLine `json:"Lines"`
-    HasOverlay  bool      `json:"HasOverlay"`
-    Message     string    `json:"Message"`
+	Lines      []OCRLine `json:"Lines"`
+	HasOverlay bool      `json:"HasOverlay"`
+	Message    string    `json:"Message"`
 }
 
 type ParsedResult struct {
-    TextOverlay           TextOverlay `json:"TextOverlay"`
-    TextOrientation       string      `json:"TextOrientation"`
-    FileParseExitCode     int         `json:"FileParseExitCode"`
-    ParsedText            string      `json:"ParsedText"`
-    ErrorMessage          string      `json:"ErrorMessage"`
-    ErrorDetails          string      `json:"ErrorDetails"`
+	TextOverlay       TextOverlay `json:"TextOverlay"`
+	TextOrientation   string      `json:"TextOrientation"`
+	FileParseExitCode int         `json:"FileParseExitCode"`
+	ParsedText        string      `json:"ParsedText"`
+	ErrorMessage      string      `json:"ErrorMessage"`
+	ErrorDetails      string      `json:"ErrorDetails"`
 }
 
 type OCRResponse struct {
-    ParsedResults                 []ParsedResult `json:"ParsedResults"`
-    OCRExitCode                   int            `json:"OCRExitCode"`
-    IsErroredOnProcessing         bool           `json:"IsErroredOnProcessing"`
-    ProcessingTimeInMilliseconds  string         `json:"ProcessingTimeInMilliseconds"`
-    SearchablePDFURL              string         `json:"SearchablePDFURL"`
+	ParsedResults                []ParsedResult `json:"ParsedResults"`
+	OCRExitCode                  int            `json:"OCRExitCode"`
+	IsErroredOnProcessing        bool           `json:"IsErroredOnProcessing"`
+	ProcessingTimeInMilliseconds string         `json:"ProcessingTimeInMilliseconds"`
+	SearchablePDFURL             string         `json:"SearchablePDFURL"`
 }
 
 func NewOCRRequest(headers map[string]string, formValues map[string]string) *OCRRequest {
